+++ webview-ui/src/App.tsx
@@ -14,7 +14,7 @@ function App() {
 
   return (
     <div>
-      <h1>Hello React!</h1>
+      <h1>Helloooo React!</h1>
       <p>Count: {count}</p>
       <button onClick={() => setCount((c) => c + 1)}>Increment</button>
       <button onClick={sendMessageToExtension}>Send VSCode Message</button>
